+++ src/tests/unit/tests/bug-filing/github/github-bug-filing-service.test.tsx
@@ -5,7 +5,7 @@ import { TextField } from 'office-ui-fabric-react/lib/TextField';
 import * as React from 'react';
 import { IMock, Mock, Times } from 'typemoq';
 
-import { createGitHubBugFilingUrl } from '../../../../../bug-filing/github/create-git-hub-bug-filing-url';
+import { createGitHubBugFilingUrl } from '../../../../../bug-filing/github/create-github-bug-filing-url';
 import { GitHubBugFilingService, GitHubBugFilingSettings } from '../../../../../bug-filing/github/github-bug-filing-service';
 import { SettingsFormProps } from '../../../../../bug-filing/types/settings-form-props';
 import { UserConfigMessageCreator } from '../../../../../common/message-creators/user-config-message-creator';
